Extract timeline generation
